+++ PlayerTrack.Plugin/Models/Models/Encounter.cs
@@ -6,7 +6,7 @@ public class Encounter
 {
     public int Id { get; set; }
 
-    public ushort TerritoryTypeId { get; set; }
+    public uint TerritoryTypeId { get; set; }
 
     public long Ended { get; set; }
