+++ PlayerTrack.Plugin/Models/Models/Player/Player.cs
@@ -47,7 +47,7 @@ public class Player : IComparable<Player>, IEquatable<Player>
 
     public uint WorldId { get; set; }
 
-    public ushort LastTerritoryType { get; set; }
+    public uint LastTerritoryType { get; set; }
 
     public ulong ContentId { get; set; }
